@@ -1,6 +1,21 @@
-import React from "react";
-import ReactDOM from "react-dom";
+import React from 'react';
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
 
-import App from "./App";
+import App from './App';
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const rootElement = document.getElementById('app');
+const root = createRoot(rootElement);
+
+root.render(
+  <StrictMode>
+    <App />
+  </StrictMode>
+);
+
+// import React from "react";
+// import ReactDOM from "react-dom";
+
+// import App from "./App";
+
+// ReactDOM.render(<App />, document.getElementById("root"));
